LAN969x: RedBox: Added decoding of Rx IFH's RED_TAG field
This is needed in order to distinguish frames redirected to the CPU because they
did not contain an HSR-tag from frames copied or redirected to the CPU for other
(good) reasons.

The ones redirected to the CPU because they didn't contain an HSR-tag can be
used by application RedBox code to raise a flag that something is wrong with the
HSR-ring.

Whether such frames should be sent to e.g. the IP-stack is still TBD. If not,
measures must be taken to filter them away before they reach it.