MEPA/MACSEC: To achieve a stable ABI, always set MEPA_MACSEC_MAX_SA to 64
The structures in the MACsec API have embedded arrays using
MEPA_MACSEC_MAX_SA to define the size. When compiling for a 1G target,
this was set to 16, when compiling for 10G target this was set to 10G.
This causes problems in designs mixing 1G and 10G PHYs as the ABI is not
compatible.

As a tempoary fix, all compile targets now uses 64 as scaling constant.

Review-record: https://bitbucket.microchip.com/projects/UNGE/repos/sw-mesa/pull-requests/293