MEPA-326 I2C read/write mapping with MEPA for M10G
MEPA/MEBA: Change the prototype of i2c read/write functions

The autogenerators (ag.rb and ag2.rb) used in this project will in some
case apply decisions based on asusmptions.

One of them is that if we have an argument called something wiht "cnt",
then the following argument must is a pointer to an array being "cnt"
long.

This was not the case with the orignial definition of the i2c read/write
APIs, and this has been changed in this commit.