MEPA: Fix indy_ts_clock_rateadj_set
When adjusting the ptp clock, it was always presumming that it was
running at 250MHz. But that is not always true because the clock can be
also 200MHz. Therefore add a check for the clock frequency and based on
this adjust with the correct value.

Review-record: https://bitbucket.microchip.com/projects/UNGE/repos/sw-mesa/pull-requests/555