laika: set RT_[ACL/ES2]_CNT_SIZE and RT_IP6PFX_CNT
RT_[ACL/ES2]_CNT_SIZE are VCAP counters that counts how many
time a rule is hit and are the same as for Laguna (No change in
this commit as the original code is a copy of Laguna).

The amount of counters available can be deduced from the amount
of VTSS_ANA_ACL_CNT_A() and VTSS_EACL_ES2_CNT() registers
available in the ANA_ACL and EACL blocks respectively.

RT_IP6PFX_CNT is the number of available fullword in the IP6PFX
VCAP. There is a single VCAP with 16 rows, which means the total
number of entries is 16. This is small compared to FireAnt and
Laguna.