Increase number of ACL rules for LMStaX builds
Increase the total number of ACL rules that can be added. Currently
increase this by 4, because we need to make sure that in LMStaX we can
add also the system rules. And not to break backwards compatibility by
taking the space from the rules the user can add, then increase the
number of rules can be added in total.