laika: add an enum for 40Gb bandwidth
Laika auto calendar can be configured for 40Gb of bandwidth on
its CPU ports. This commit adds a new entry to
vtss_internal_bw_t and mesa_internal_bw_t.

The enum (fa_cal_speed_t in vtss_fa.c) used to represent 40G in
hardware (in the VTSS_QSYS_CAL_AUTO register) is 5, which
conflicts with the encoding of 25G for Sparx5 and Laguna.

The hardware designer had to reuse the value 5 for 40G since
this field is encoded on 3 bits which means 7 is the max value.
6 and 7 are not in the enum but represent 0.5G and 1.5G
respectively in hardware. Since 25G isn't used on laika, 5 has
been repurposed to represent 40G or whatever the maximum
bandwidth is on that arch [1]. For that reason, this commit
renames FA_CAL_SPEED_25G into FA_CAL_SPEED_MAX and introduce
some arch specific logic to convert to 25G or 40G.

[1]: rtl-switch 26d6ab90cc9 ("Most of Laika seems to work.")
file src/design/vit_gz_qsys_lib/blk/cbc/cbc-rtl.arc