MEBA: Ocelot: Enable back the DEV_ALL interrupts
On Ocelot platform it was noticed that once a cable was plugged into one
of the internal PHYs, the interrupt from internal PHY was generated and
was disabled to handle it, it was handled, but it was never enabled
back. Meaning that when the link was removed or plug back the interrupt
was not generated again because it was never enabled.
This issue has not been seen because we have the poling mechanism so
after 1 second, it would detect anyway if the link is up or down.
But this can be seen if you map the interrupt on a GPIO and check there.
The fix consists in enabling back the interrupt after was handled.