MESA-1019. LAN969x/SparX-5. Disabling disassembler's underrun watchdog.
Boards with few ports per taxi bus get higher bandwidth than boards with many
ports per taxi bus. This higher bandwidth may drain the disassembler's (DSM's)
FIFO, causing underrun and a subsequent abortion of the frame, so that it never
comes out. This is in particular the case if enabling the RedBox on the Taxi
bus, because it has a larger FIFO that can drain the DSM without push-back. The
solution is to disable the DSM's underrun watchdog, so that it cannot abort
frames due to underrun.