add support for CPU masquerading
In order to forward packets to a CPU port in a way that it
behaves like a front port, a CPU masquerading mechanism is
introduced.

CPU masquerading steals a physical front port and redirect all
its traffic to a CPU port and uses the physical port number to
inject frames.