laika: define PSFP constants
VTSS_JR_SDX_CNT is FFL_CHIP_NUM_SDX minus 1, which is:
2048 - 1 = 2047.

VTSS_JR_POL_CNT is the same as Laguna (256). It is equal to
"FFL_ANA_NUM_SDLB - 8*FFL_CHIP_NUM_ETH_PORTS" which is:
512 - 8*32 = 256

VTSS_JR_STAT_CNT is FFL_CHIP_NUM_SERVICE_STAT which is 2048

The code previously used the constant values of sparx5 which
caused the recent modifications for HQoS to cause illegal
instruction accesses on Laika.