cpu masquerading: fix null classified VID
Packets injected with a TOFH have their TOFH translated to an IFH when
entering the ASM. This IFH use ANA_CLM as its pipeline injection point
which means that basic classification is bypassed. This IFH is created
with a classified VID of 0 which doesn't get fixed by VLAN
classification since it is bypassed.

Fix this using the default CLM-A[0] action if the port is a masqueraded
port.