vtss_fa_packet: Update the extractions of the frame
If a frame with a length of 1518 is reaching the CPU then the lmstax
will drop the frame saying that the buffer is too short even if it has
length of 1518. The problem is that if the logic extracting the frame is
using 4 byte and for a frame which has 1518 bytes which is not a
multiple of 4 will say that it doesn't have more space last 2 byte and
it would drop the frame but it still has.
Fix this by checking if there is also space for 2 bytes.